projeto-3/requirements.txt
boto3==1.28.73
botocore==1.31.73
certifi==2023.7.22
charset-normalizer==3.3.1
idna==3.4
jmespath==1.0.1
python-dateutil==2.8.2
requests==2.31.0
s3transfer==0.7.0
six==1.16.0
urllib3==2.0.7
